refactor: update hooks filename for cart, product and favorite
