+++ src/pages/Home.tsx
@@ -32,7 +32,7 @@ import { categoryAtom } from "../atoms/products";
 import { chevronForwardOutline } from "ionicons/icons";
 import { getAuth } from "firebase/auth";
 import { useCollectionDataOnce } from "react-firebase-hooks/firestore";
-import useFavorite from "../hooks/favorite";
+import useFavorite from "../hooks/useFavorite";
 import { useRefresh } from "../hooks/page";
 import { useSetRecoilState } from "recoil";
